Refactor types to separate file
